Add ability to supply custom threat file
This change adds a new threat feed configuration option `customThreatFile`. The option allows you to specify a text file containing additional content for the threat feed. The contents of the custom threat file are appended to the output of the threat feed. This could be IP addresses or any other type of text.