Change banner config to allow multiple headers
This update modifies handling of the 'banner' configuration for HTTP/S honeypots. Previously, the 'banner' string was used to set the "Server" header in HTTP responses. Now, the configuration takes a comma-separated string of key/value pairs, allowing the specification of multiple custom headers. For example, the string "Server: IIS/8.5, X-Powered-By: ASP.NET" could be used to set both the "Server" header and "X-Powered-By" header.