Add ability to define rules for when a request updates the threat feed
This change adds Rule and Rules structs for HTTP honeypot configurations. The rules are regex patterns that define when an HTTP request should trigger an update the threat feed.

The Rule struct allows you specify a target that defines which part of the HTTP request should match your pattern.

Example configuration:
```
<server type="http">
  <rules>
    <match target="path">\.php$</match>
  </rules>
</server>
```

This example triggers an update to the threat feed only if the HTTP request path ends with `.php`. Any other request will not trigger an update.