Fix: Don't log "User-Agent" in `headers` key
This commit ensures the `User-Agent` value is removed from the HTTP request headers when logging the request. `User-Agent` is logged in the `event_details` array, so it should not duplicate the value in the `headers` array when logging. This commit fixes the issue introduced in commit 12ada38faad6fa5172ba54ceec7fccc1d5eca829 which normalised header names to lowercase.